TOKEN TOOL/requirements.txt
flask==3.0.0
requests==2.31.0
pycryptodome==3.19.0
pyotp==2.9.0
gunicorn==21.2.0
